APIs/azure.com/subscription-subscriptionDefinitions/2017-11-01-preview/python-aiohttp-client/tests/test_subscription_definitions_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.error_response import ErrorResponse
from openapi_server.models.operation_list_result import OperationListResult
from openapi_server.models.subscription_definition import SubscriptionDefinition
from openapi_server.models.subscription_definition_list import SubscriptionDefinitionList


pytestmark = pytest.mark.asyncio

async def test_subscription_definitions_create(client):
    """Test case for subscription_definitions_create

    
    """
    body = {"name":"name","id":"id","type":"type","properties":{"offerType":"offerType","etag":"etag","subscriptionId":"subscriptionId","subscriptionDisplayName":"subscriptionDisplayName"}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/providers/Microsoft.Subscription/subscriptionDefinitions/{subscription_definition_name}'.format(subscription_definition_name='subscription_definition_name_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_subscription_definitions_get(client):
    """Test case for subscription_definitions_get

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/providers/Microsoft.Subscription/subscriptionDefinitions/{subscription_definition_name}'.format(subscription_definition_name='subscription_definition_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_subscription_definitions_get_operation_status(client):
    """Test case for subscription_definitions_get_operation_status

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/providers/Microsoft.Subscription/subscriptionOperations/{operation_id}'.format(operation_id='operation_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_subscription_definitions_list(client):
    """Test case for subscription_definitions_list

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/providers/Microsoft.Subscription/subscriptionDefinitions',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_subscription_definitions_operation_metadata_list(client):
    """Test case for subscription_definitions_operation_metadata_list

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/providers/Microsoft.Subscription/operations',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

